--- dea_waterbodies/__init__.py
@@ -1,1 +1,0 @@
-from dea_waterbodies.__version__ import __version__  # noqa: F401
