+++ deafrica_waterbodies/__init__.py
@@ -1,0 +1,1 @@
+from deafrica_waterbodies.__version__ import __version__  # noqa: F401
